Renamed NSCFLTimeStepSize to CFLTimeStepSize
Now that the THM version no longer exists, the correct name can
be used.

Closes #20265
