+++ framework/doc/content/source/timesteppers/PostprocessorDT.md
@@ -3,7 +3,7 @@
 !syntax description /Executioner/TimeStepper/PostprocessorDT
 
 The postprocessor can be computing a global metric, like the Courant Friedrich Levy criterion
-with the [NSCFLTimeStepSize](NSCFLTimeStepSize.md optional=True) or the
+with the [CFLTimeStepSize](CFLTimeStepSize.md optional=True) or the
 [LevelSetCFLCondition](LevelSetCFLCondition.md optional=true) postprocessor. In those cases,
 the postprocessor computes the maximum value of the timestep that should ensure stability.
 Variations in numerical schemes may impact the accuracy of this postprocessor, so the
