+++ modules/navier_stokes/include/base/NS.h
@@ -158,5 +158,3 @@ namespace NS_CONSTANTS
 {
 using namespace HeatConduction::Constants;
 }
-
-#define NS_HAS_CFL_TIME_STEP_SIZE 1
